qemu: support 64-bit latency histogram boundaries
QEMU uses uint64 values for block latency histogram boundaries, while
libvirt currently stores and parses them as unsigned int values.

This limits histogram boundaries to UINT32_MAX nanoseconds and prevents
using larger values supported by QEMU.

Use unsigned long long for latency histogram boundaries throughout the
domain configuration and QEMU monitor code. Update the XML schema and
parser accordingly, and serialize the values using the 64-bit JSON
number helper.

Extend the XML and QEMU monitor tests with a 5000000000 nanosecond
boundary to verify values above UINT32_MAX are preserved.

Resolves: https://gitlab.com/libvirt/libvirt/-/work_items/919

+++ src/conf/domain_conf.h
@@ -599,11 +599,11 @@ struct _virDomainDiskDef {
     unsigned int *statistics; /* Optional, zero terminated list of intervals to
                                 collect statistics for */
     /* optional zero terminated lists of bin boundaries for latency histograms */
-    unsigned int *histogram_boundaries;
-    unsigned int *histogram_boundaries_read;
-    unsigned int *histogram_boundaries_write;
-    unsigned int *histogram_boundaries_zone;
-    unsigned int *histogram_boundaries_flush;
+    unsigned long long *histogram_boundaries;
+    unsigned long long *histogram_boundaries_read;
+    unsigned long long *histogram_boundaries_write;
+    unsigned long long *histogram_boundaries_zone;
+    unsigned long long *histogram_boundaries_flush;
 
     virDomainDiskDetectZeroes detect_zeroes;
     virTristateSwitch discard_no_unref;
