+++ src/qemu/qemu_monitor.c
@@ -4558,11 +4558,11 @@ qemuMonitorBlockdevSetActive(qemuMonitor *mon,
 int
 qemuMonitorBlockLatencyHistogramSet(qemuMonitor *mon,
                                     const char *id,
-                                    unsigned int *boundaries,
-                                    unsigned int *boundaries_read,
-                                    unsigned int *boundaries_write,
-                                    unsigned int *boundaries_zone,
-                                    unsigned int *boundaries_flush)
+                                    unsigned long long *boundaries,
+                                    unsigned long long *boundaries_read,
+                                    unsigned long long *boundaries_write,
+                                    unsigned long long *boundaries_zone,
+                                    unsigned long long *boundaries_flush)
 {
     QEMU_CHECK_MONITOR(mon);
     VIR_DEBUG("id='%s'", id);
